+++ R/ErrorMessages.R
@@ -118,28 +118,6 @@
 
 
 
-..warning_geom_not_available <- function(
-  x,
-  call = rlang::caller_env()
-) {
-  message_string <- paste0(
-    "The following Geom object was not available: ", x, " ",
-    "This may occur if ggplot2 was not available when familiar was first loaded ",
-    "(e.g., library(familiar)), but subsequently installed or added to library search paths. ",
-    "To prevent this issue, ensure that ggplot2 is installed prior to loading familiar."
-  )
-  
-  rlang::warn(
-    message = message_string,
-    class = c("familiar_warning", "plot_warning"),
-    call = call
-  )
-  
-  return(invisible(TRUE))
-}
-
-
-
 ..deprecation_count <- function(as_error = FALSE) {
   
   message_string <- "The \"count\" outcome type has been deprecated in familiar version 2.0.0."