+++ NEWS.md
@@ -22,6 +22,9 @@
 - If `ggplot2` is not installed or made visible on library paths prior to loading `familiar`,
   several `Geom` objects, used for plotting SHAP waterfall and force plots, could not be created.
   This now causes a warning instead of an error.
+  
+- Fixed an issue that occurred when parsing `n_important_features` when explicitly
+  expressing the corresponding data elements.
 
 # Version 2.0.2 (Cackling Caribou)